// apps/admin-web/src/api/model/profileControllerSetup2FA200.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Admin API
 * The Admin API description
 * OpenAPI spec version: 1.0
 */
import type { BaseResponse } from './baseResponse';
import type { Setup2FAResponseDto } from './setup2FAResponseDto';

export type ProfileControllerSetup2FA200 = BaseResponse & {
  data: Setup2FAResponseDto;
};