// apps/admin-web/src/api/model/termAgreementResponseDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Admin API
 * The Admin API description
 * OpenAPI spec version: 1.0
 */
import type { Organization } from './organization';
import type { TermAgreementResponseDtoMetadata } from './termAgreementResponseDtoMetadata';
import type { TermAgreementResponseDtoTermType } from './termAgreementResponseDtoTermType';
import type { TermCategory } from './termCategory';

export interface TermAgreementResponseDto {
  category: TermCategory;
  content: string;
  version: string;
  startDate: string;
  endDate: string;
  organization?: Organization;
  termType: TermAgreementResponseDtoTermType;
  agreementId?: string;
  id: string;
  createdAt: string;
  createdBy?: string;
  updatedAt: string;
  updatedBy?: string;
  deletedAt?: string;
  deletedBy?: string;
  metadata?: TermAgreementResponseDtoMetadata;
  isActive: boolean;
  title: string;
  agreedAt?: string;
}